Updates for calling by id and supporting future view details of non-draft program profiles
