+++ src/ECER.Clients.PSPPortal/ecer.clients.pspportal.client/src/store/program.ts
@@ -35,21 +35,11 @@ export const useProgramStore = defineStore("program", {
     }
   },
   actions: {
-    async setDraftProgram(programId: string) {
-      // Drop any existing draft program
-      this.$reset();
-
-      const { data: programs } = await getPrograms(programId);
-      const program = programs && programs.length > 0 ? programs[0] : undefined;
-
-      if (program) {
-        if(!program.portalStage) {
-          program.portalStage = "ProgramOverview";
-        }
-        this.draftProgram = program;
+    setDraftProgramFromProfile(program: Components.Schemas.Program): void {
+      if (!program.portalStage) {
+        program.portalStage = "ProgramOverview";
       }
-
-      return programs;
+      this.draftProgram = program;
     },
     prepareDraftProgramFromWizard() {
       const wizardStore = useWizardStore();
